split tests into multiple files.
