+++ tests/core_tests/test_api.py
@@ -9,23 +9,20 @@
 # not expressly granted therein are reserved by Shotgun Software Inc.
 
 import os
-import sys
-import unittest2 as unittest
 
 from mock import Mock, patch
-from tank_vendor import yaml
 
-import sgtk
 
 import tank
 from tank.api import Tank
-from tank.errors import TankError
 from tank.template import TemplatePath, TemplateString
 from tank.templatekey import StringKey, IntegerKey, SequenceKey
 
-from tank_test.tank_test_base import *
+from tank_test.tank_test_base import TankTestBase, setUpModule # noqa
+
 
 class TestInit(TankTestBase):
+    """Tests basic initialization of the sgtk API"""
 
     def setUp(self):
         super(TestInit, self).setUp()
@@ -489,347 +486,6 @@ class TestApiCache(TankTestBase):
         self.assertEquals(tk.get_cache_item("foo"), None)
         self.assertEquals(tk2.get_cache_item("foo"), None)
 
-
-class TestTankFromPath(TankTestBase):
-
-    def setUp(self):
-        super(TestTankFromPath, self).setUp()
-        self.setup_multi_root_fixtures()
-
-    def test_primary_branch(self):
-        """
-        Test path from primary branch.
-        """
-        child_path = os.path.join(self.project_root, "child_dir")
-        os.mkdir(os.path.join(self.project_root, "child_dir"))
-        result = tank.tank_from_path(child_path)
-        self.assertIsInstance(result, Tank)
-        self.assertEquals(result.project_path, self.project_root)
-
-    def test_alternate_branch(self):
-        """
-        Test path not from primary branch.
-        """
-        os.mkdir(os.path.join(self.alt_root_1, "child_dir"))
-        child_path = os.path.join(self.alt_root_1, "child_dir")
-        result = tank.tank_from_path(child_path)
-        self.assertIsInstance(result, Tank)
-        self.assertEquals(result.project_path, self.project_root)
-
-    def test_bad_path(self):
-        """
-        Test path not in project tree.
-        """
-        bad_path = os.path.dirname(self.tank_temp)
-        self.assertRaises(TankError, tank.tank_from_path, bad_path)
-
-    def test_tank_temp(self):
-        """
-        Test passing in studio path.
-        """
-        self.assertRaises(TankError, tank.tank_from_path, self.tank_temp)
-
-
-class TestTankFromPathDuplicatePcPaths(TankTestBase):
-    """
-    Test behavior and error messages when multiple pipeline
-    configurations are pointing at the same location
-    """
-
-    def setUp(self):
-        super(TestTankFromPathDuplicatePcPaths, self).setUp()
-
-        # define an additional pipeline config with overlapping paths
-        self.overlapping_pc = {
-            "type": "PipelineConfiguration",
-            "code": "Primary",
-            "id": 123456,
-            "project": self.project,
-            "windows_path": self.pipeline_config_root,
-            "mac_path": self.pipeline_config_root,
-            "linux_path": self.pipeline_config_root}
-
-        self.add_to_sg_mock_db(self.overlapping_pc)
-
-    def test_primary_duplicates_from_path(self):
-        """
-        Test primary dupes
-        """
-        self.assertRaisesRegexp(TankError,
-                                "The path '.*' is associated with more than one Primary pipeline configuration.",
-                                sgtk.sgtk_from_path,
-                                self.project_root)
-
-    def test_primary_duplicates_from_entity(self):
-        """
-        Test primary dupes
-        """
-        self.assertRaisesRegexp(TankError,
-                                "More than one primary pipeline configuration is associated with the entity",
-                                sgtk.sgtk_from_entity,
-                                "Project",
-                                self.project["id"])
-
-class TestTankFromEntityWithMixedSlashes(TankTestBase):
-    """
-    Tests the case where a Windows local storage uses forward slashes.
-    """
-
-    def test_with_mixed_slashes(self):
-        """
-        Check that a sgtk init works for this path
-        """
-        # only run this test on windows
-        if sys.platform == "win32":
-
-            self.sg_pc_entity["windows_path"] = self.pipeline_config_root.replace("\\", "/")
-            self.add_to_sg_mock_db(self.sg_pc_entity)
-            self.add_to_sg_mock_db(self.project)
-            self.add_to_sg_mock_db({
-                "type": "Shot",
-                "id": 1,
-                "project": self.project
-            })
-
-            os.environ["TANK_CURRENT_PC"] = self.pipeline_config_root
-            try:
-                sgtk.sgtk_from_entity("Shot", 1)
-            finally:
-                del os.environ["TANK_CURRENT_PC"]
-
-
-class TestGetConfigInstallLocationPathSlashes(TankTestBase):
-    """
-    Tests the case where a Windows config location uses double slashes.
-    """
-
-    @patch("tank.pipelineconfig_utils._get_install_locations")
-    def test_config_path_cleanup(self, get_install_locations_mock):
-        """
-        Check that any glitches in the path are correctly cleaned up.
-        """
-        # only run this test on windows
-        if sys.platform == "win32":
-
-            # This path has multiple issues we've encountered in the wild
-            # It without any escaping sequence, it reads as
-            # "   C:/configs\\site//project   "
-            # 1. it has whitespace at the begging and end.
-            # 2. Uses double slashes instead of single backslash
-            # 3. Uses slash instead of backslash
-            # 4. Uses double backslashes as a folder separator.
-            get_install_locations_mock.return_value = sgtk.util.ShotgunPath("   C:/configs\\\\site//project   ", None, None)
-
-            self.assertEqual(
-                # We don't need to pass an actual path since _get_install_location is mocked.
-                tank.pipelineconfig_utils.get_config_install_location(None),
-                "C:\\configs\\site\\project"
-            )
-
-
-class TestTankFromPathWindowsNoSlash(TankTestBase):
-    """
-    Tests the edge case where a Windows local storage is set to be 'C:'
-    """
-
-    PROJECT_NAME = "temp"
-    STORAGE_ROOT = "C:"
-
-    def setUp(self):
-
-        # set up a project named temp, so that it will end up in c:\temp
-        super(TestTankFromPathWindowsNoSlash, self).setUp(parameters = {"project_tank_name": self.PROJECT_NAME})
-        
-        # set up std fixtures
-        self.setup_fixtures()
-
-        # patch primary local storage def
-        self.primary_storage["windows_path"] = self.STORAGE_ROOT
-        # re-add it
-        self.add_to_sg_mock_db(self.primary_storage)
-
-        # now re-write roots.yml
-        roots = {"primary": {}}
-        for os_name in ["windows_path", "linux_path", "mac_path"]:
-            #TODO make os specific roots
-            roots["primary"][os_name] = self.sg_pc_entity[os_name]
-        roots_path = os.path.join(self.pipeline_config_root,
-                                  "config",
-                                  "core",
-                                  "roots.yml")
-        roots_file = open(roots_path, "w")
-        roots_file.write(yaml.dump(roots))
-        roots_file.close()
-
-        # need a new pipeline config object that is
-        # using the new roots def file we just created
-        self.pipeline_configuration = sgtk.pipelineconfig_factory.from_path(self.pipeline_config_root)
-        # push this new pipeline config into the tk api
-        self.tk._Tank__pipeline_config = self.pipeline_configuration
-        # force reload templates
-        self.tk.reload_templates()
-
-
-    def test_project_path_lookup(self):
-        """
-        Check that a sgtk init works for this path
-        """
-        # only run this test on windows
-        if sys.platform == "win32":
-
-            # probe a path inside of project
-            test_path = "%s\\%s\\toolkit_test_path" % (self.STORAGE_ROOT, self.PROJECT_NAME)
-            if not os.path.exists(test_path):
-                os.makedirs(test_path)
-            self.assertIsInstance(sgtk.sgtk_from_path(test_path), Tank)
-
-
-
-
-
-class TestTankFromPathOverlapStorage(TankTestBase):
-    """
-    Tests edge case with overlapping storages
-
-    For example, imagine the following setup:
-    Storages: f:\ and f:\foo
-    Project names: foo and bar
-    This means we have the following project roots:
-    (1) f:\foo      (storage f:\, project foo)
-    (2) f:\bar      (storage f:\, project bar)
-    (3) f:\foo\foo  (storage f:\foo, project foo)
-    (4) f:\foo\bar  (storage f:\foo, project bar)
-
-    The path f:\foo\bar\hello_world.ma could either belong to
-    project bar (matching 4) or project foo (matching 1).
-
-    In this case, sgtk_from_path() should succeed in case you are using a local
-    tank command or API and fail if you are using a studio level command.
-
-    """
-
-    def setUp(self):
-
-        # set up two storages and two projects
-        super(TestTankFromPathOverlapStorage, self).setUp(parameters = {"project_tank_name": "foo"})
-
-        # add second project
-        self.project_2 = {"type": "Project",
-                          "id": 2345,
-                          "tank_name": "bar",
-                          "name": "project_name"}
-
-        # define entity for pipeline configuration
-        self.project_2_pc = {"type": "PipelineConfiguration",
-                             "code": "Primary",
-                             "id": 123456,
-                             "project": self.project_2,
-                             "windows_path": "F:\\temp\\bar_pc",
-                             "mac_path": "/tmp/bar_pc",
-                             "linux_path": "/tmp/bar_pc"}
-
-        self.add_to_sg_mock_db(self.project_2)
-        self.add_to_sg_mock_db(self.project_2_pc)
-
-        # set up std fixtures
-        self.setup_multi_root_fixtures()
-
-        # patch storages
-        self.alt_storage_1["windows_path"] = "C:\\temp"
-        self.alt_storage_1["mac_path"] = "/tmp"
-        self.alt_storage_1["linux_path"] = "/tmp"
-
-        self.alt_storage_2["windows_path"] = "C:\\temp\\foo"
-        self.alt_storage_2["mac_path"] = "/tmp/foo"
-        self.alt_storage_2["linux_path"] = "/tmp/foo"
-
-        self.add_to_sg_mock_db(self.alt_storage_1)
-        self.add_to_sg_mock_db(self.alt_storage_2)
-
-        # Write roots file
-        roots = {"primary": {}, "alternate_1": {}, "alternate_2": {}}
-        for os_name in ["windows_path", "linux_path", "mac_path"]:
-            roots["primary"][os_name] = os.path.dirname(self.project_root)
-            roots["alternate_1"][os_name] = self.alt_storage_1[os_name]
-            roots["alternate_2"][os_name] = self.alt_storage_2[os_name]
-        roots_path = os.path.join(self.pipeline_config_root, "config", "core", "roots.yml")
-        roots_file = open(roots_path, "w")
-        roots_file.write(yaml.dump(roots))
-        roots_file.close()
-
-        # need a new pipeline config object that is using the new
-        # roots def file we just created
-        self.pipeline_configuration = sgtk.pipelineconfig_factory.from_path(self.pipeline_config_root)
-        # push this new pipeline config into the tk api
-        self.tk._Tank__pipeline_config = self.pipeline_configuration
-        # force reload templates
-        self.tk.reload_templates()
-
-
-    def test_project_path_lookup_studio_mode(self):
-        """
-        When running this edge case from a studio install, we expect an error:
-
-        TankError: The path '/tmp/foo/bar' is potentially associated with more than one primary
-        pipeline configuration. This can happen if there is ambiguity in your project setup,
-        where projects store their data in an overlapping fashion. In this case, try creating
-        your API instance (or tank command) directly from the pipeline configuration rather
-        than via the studio level API. This will explicitly call out which project you are
-        intending to use in conjunction with he path. The pipeline configuration paths
-        associated with this path are:
-        ['/var/folders/fq/65bs7wwx3mz7jdsh4vxm34xc0000gn/T/tankTemporaryTestData_1422967258.765262/pipeline_configuration',
-        '/tmp/bar_pc']
-
-        """
-
-        probe_path = {}
-        probe_path["win32"] = "C:\\temp\\foo\\bar\\test.ma"
-        probe_path["darwin"] = "/tmp/foo/bar/test.ma"
-        probe_path["linux2"] = "/tmp/foo/bar/test.ma"
-
-        test_path = probe_path[sys.platform]
-        test_path_dir = os.path.dirname(test_path)
-
-        if not os.path.exists(test_path_dir):
-            os.makedirs(test_path_dir)
-
-        self.assertRaisesRegexp(TankError,
-                                "The path '.*' is associated with more than one Primary pipeline configuration.",
-                                sgtk.sgtk_from_path,
-                                test_path)
-
-
-
-    def test_project_path_lookup_local_mode(self):
-        """
-        Check that a sgtk init works for this path
-        """
-
-        # By setting the TANK_CURRENT_PC, we emulate the behaviour
-        # of a local API running. Push this variable
-        old_tank_current_pc = None
-        if "TANK_CURRENT_PC" in os.environ:
-            old_tank_current_pc = os.environ["TANK_CURRENT_PC"]
-        os.environ["TANK_CURRENT_PC"] = self.pipeline_config_root
-
-        probe_path = {}
-        probe_path["win32"] = "C:\\temp\\foo\\bar\\test.ma"
-        probe_path["darwin"] = "/tmp/foo/bar/test.ma"
-        probe_path["linux2"] = "/tmp/foo/bar/test.ma"
-
-        test_path = probe_path[sys.platform]
-        test_path_dir = os.path.dirname(test_path)
-
-        if not os.path.exists(test_path_dir):
-            os.makedirs(test_path_dir)
-
-        self.assertIsInstance(sgtk.sgtk_from_path(test_path), Tank)
-
-        # and pop the modification
-        if old_tank_current_pc is None:
-            del os.environ["TANK_CURRENT_PC"]
-        else:
-            os.environ["TANK_CURRENT_PC"] = old_tank_current_pc
